i2c_mspm0g3xxx: Serialize i2c-target handling
As the driver is not context aware of the i2c-target handles, it might not
be the best idea to call into these from the ISR itself.
The callbacks might try to lock the thread causing severe issues if called
from ISR.

By offloading the target handling we ensure that the ISR is executed
swiftly and is merely a signal trigger for what's happening next.
As clock stretching is enforced, this should not cause any clock issues.

This commit also adds a kconfig to enable the target handles.
Now we're spawning a worker thread why the overhead is too big if not utilized.

Also the zephyr api will mark the i2c-target callbacks as not implemented if
attempted to call without the flag.

Signed-off-by: Hans Binderup <habi@bang-olufsen.dk>