i2c_mspm0g3xxx: don't disable i2c-target mode when registering new targets
Disabling i2c-target mode for the i2c reg would cause the clock to get
stuck. Could cause a lot of funky interactions.

As the state model is protected by a semaphore and we're already in control
of the isr + target is already registered, this should be safe to do.

Signed-off-by: Hans Binderup <habi@bang-olufsen.dk>