i2c_mspm0g3xxx: add "target-mode-only" property
Setting to disable initial config of i2c-controller.
As the initial i2c-controller mode can cause clock issues when
another clock controller is present on the bus, it can be useful to disable
the initial setup.

Also the peripheral will not enable controller mode when no targets are
registered.

Ie. the bus will only be usable for targets.

Signed-off-by: Hans Binderup <habi@bang-olufsen.dk>