i2c_mspm0g3xxx: scl stuck: disable target wakeup
Based on TI's comment:

It turns out that we have an errata with the target low-power wakeup
feature, which can cause the target to hold the SCL line low indefinitely.

There is a logical error within the target wake up enable feature controlled
by the bit SCTR.SWUEN which has the ability to stretch the Clock Signal
indefinitely when experiencing unexpected traffic on the bus (ex. some
other transaction). Thus it is recommended to turn off the feature by
writing 0 to this bit after the device emerges from a reset.

After multiple 10+ min tests, i could not reproduce the issue

Signed-off-by: Dimitris Karnikis <dika@bang-olufsen.dk>