drivers: counter: mspm0 add hw timers drivers
This pr introduces a new device driver to use some of the mspm0 hardware
timers. More specifically TimerG{0, 6, 7, 12}. There are more timers
available but they need to be configured differently, so for the time
being, i decided to keep it simple for our use cases. More specifically,
we can now set the following properties for a timer (in dts):

1. prescale
2. divide-ration
3. resolution
4. mode (ONE_SHOT/PERIODIC/other combinations)

The period has to be configured from the application part, as we are
leveraging Zephyr's counter timer API. We are passing the period (ticks)
as part of the configuration for the counter alongside with the callback.
Also, several of the timers support multiple channels, but are configured
to be in different modes that would make things more complicated. Also,
they are also other modes/properties available that are tied to specific
timers, but we could also add them on a future version. MAX_CHANNELS
are hardcoded to 1, so some parts of the counter API are kind of unused.

Example: to use TimerG12 to run a periodic timer every 1s

```
/*
 * Timer clock configuration to be sourced by BUSCLK /  (32000000 Hz)
 * timerClkFreq = (timerClkSrc / (timerClkDivRatio * (timerClkPrescale + 1)))
 *   32000000 Hz = 32000000 Hz / (1 * (0 + 1))
 * timerPeriod = (32000000 Hz) - 1
 */

&timg12 {
    status = "okay";
    compatible = "ti,mspm0-counter";
};

const struct device *dev0 = DEVICE_DT_GET(DT_NODELABEL(timg12));
alarm_cfg.ticks = counter_us_to_ticks(dev0, 1000000);
alarm_cfg.callback = my_isr;
alarm_cfg.user_data = &alarm_cfg;
counter_set_channel_alarm(dev0, 0, &alarm_cfg);
counter_start(dev0);
```

Co-authored-by: Hans Binderup <32706408+hansbinderup@users.noreply.github.com>
Signed-off-by: Dimitris Karnikis <dika@bang-olufsen.dk>