drivers: serial: fix uart_mspm0 race condition
Previously the callbacks were dependent on the actual interrupt register.
This is an issue as it's prone to change before handling the rx/tx
callbacks. If this happened we would not clear given interrupt status (from
the ISR) and the UART would end up completely hanging.

Instead use the provided serialization handle to get the interrupt with the
highest priority, handle given interrupt, repeat.

Signed-off-by: Hans Binderup <habi@bang-olufsen.dk>