drivers: i2c: mspm0: add transfer timeout
Previously the controller would be stuck in a busy loop waiting for the
target to respond.
If the target is not attached it would mean that we'd end up in a
situation where the busy loop would wait indefinite.
Also in general using a busy wait is performance heavy where we can
offload the wait using sem synchronization.

Signed-off-by: Hans Binderup <habi@bang-olufsen.dk>