drivers: i2c: mspm0: correct controller fifo handling on error
Prior to this commit the controller seemed to be corrupted whenever an
error occurred.

This could be caused in a few ways:
* Peripheral not available (fifo never transmitted)
* Timeout handling (fifo corrupted)

Now the controller will still work even if a peripheral has not
responded or is not available. No reason to drag down the entire bus
just because a single peripheral is not available.

Signed-off-by: Hans Binderup <habi@bang-olufsen.dk>