drivers: pwm: mspm0 add pwm timers
This pr introduces a new device driver to use pwm functionality on
some of the available TIMA/TIMG. In the dts file, we can define multiple
properties such as:

1. pinctrl
2. clock devide ratio
3. prescaler (if supported)
4. pwm mode (check docs)
5. capture compare channel index and channel direction

and more.

The period has to be configured from the application part, as we are
leveraging Zephyr's pwm driver API. Also, several of the timers support
multiple channels, but are configured to be in different modes that would
make things more complicated.

pwm_mspm0_set_cycles expect period_cycles, so the calculation has to take
place in the application code

Signed-off-by: Dimitris Karnikis <dika@bang-olufsen.dk>
Co-authored-by: Emil-Juhl <59731040+Emil-Juhl@users.noreply.github.com>