i2c: mspm0g3xxx: Implement repeated start
Now we inject the stop condition only on buffers that are flagged as
stop. This doesn't follow the I2C_MSG_REPEAT bit, this is because the
bit may be in the following buffer, making the changes needed more
complex. If we need more complex behavior, we will need to further
modify the driver.