mspm0g3xxx: Add option to enable hfxt for CAN
We need to be able to run CAN using an external clock source, since the
in chip package is not up to spec. TI had already added a way to
configure the clock pins via macros. I am just adding a config option so
we can enable this on specific boards needing the extra feature