can: Fix the enablement of HFXT
We use to have custom settings for enabling HFXT in soc.c which
disappeared on new TI 4.0.99 branch. Thus letting CAN without
the possibility to use HXFT.

It is now reinstalled:
- use CONFIG_CAN_USE_HFXT to enable it or not in clock_control driver
- use CONFIG_CAN_USE_HFXT in can versus the old SOC_MSPM0_CAN_USE_HFXT
which was confusing. A Kconfig option is better anyway.

Unlike the former branch, we _do not_ disable syspll which is now used
by a few driver to drive their inner clock. Before all were relying on
the base sysclk. And since according to datasheet, it is possible to
enable both syspll and HFXT, let's do so relevantly.

Important note: this is a q&d fix. A proper fix would be to introduce
a hfxt clock node for mspm0 parent clock, add a frequency property and
enable/disable hfxt depending on that new node being enabled or not
while setting the right frequency as well (the overall range is between
4 to 40 MHz. We use 38-40Mhz as a default becauses it suits us, but
it is not generic).

Signed-off-by: Tomasz Bursztyka <tobu@bang-olufsen.dk>