drivers: i2c: i2c_mspm0: make thread toggable
Give the option to select thread support or not by defining:
CONFIG_I2C_MSPM0_THREAD_SUPPORT

It is observed that the i2c operations sometime hang. Our suspicion is
that the i2c thread is getting blocked for some reason. To mitigate
that without breaking any functionality, try to make the i2c thread
configurable.

If thread support is not enabled, all the i2c event handling will be
done in the same thread(during ISR) and will be served at the same
time. Of course, this might cause side effects with the current
drivers; we would need to spend some time investigating.

Signed-off-by: Dimitris Karnikis <dika@bang-olufsen.dk>