fixup! drivers: Add led current setting support on lp5569
Fix the devicetree parsing by replacing dashes (-) with underscores (_) as
is required in zephyr.

Fix the node lookup for the led_max_microamp property by using
DT_INST_PROP_OR. The `id` that is passed along doesn't have the format that
DT_PROP_OR requires and thus the lookup would always fail and the default
value would be used.

Signed-off-by: Emil Dahl Juhl <emdj@bang-olufsen.dk>